pyproject.toml
[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[project]
name = "langfuse-utils"
version = "0.1.0"
description = "A collection of utilities for Langfuse"
readme = "README.md"
requires-python = ">=3.14"
license = { text = "MIT" }
dependencies = [
    "httpx>=0.28.1",
    "rich>=14.2.0",
    "typer>=0.20.0",
    "python-dotenv>=1.0.1",
]

[project.scripts]
langfuse-utils = "langfuse_utils.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["copy_traces", "langfuse_utils"]

[dependency-groups]
dev = [
    "poethepoet>=0.38.0",
    "pre-commit>=4.5.0",
    "pymarkdownlnt>=0.9.33",
    "pytest>=9.0.1",
    "pytest-cov>=7.0.0",
    "pytest-xdist>=3.8.0",
    "ruff>=0.14.6",
]

[tool.pytest.ini_options]
addopts = "--cov=copy_traces --cov=langfuse_utils --cov-branch --cov-fail-under=100"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py314"
lint.select = ["E", "F", "I", "B"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false

[tool.poe.tasks]
ruff = "ruff check ."
ruff-fmt = "ruff format --check ."
pytest = "pytest -n auto"
pre-commit = "pre-commit run --all-files"
markdown = "pymarkdown --strict-config --config pymarkdown.json scan ."
check = ["ruff", "ruff-fmt", "pytest", "pre-commit", "markdown"]
